requirements.txt
flask-cors==4.0.0
Flask==2.3.0
python-dotenv==1.0.0
openai==1.5.0
httpx==0.25.0
requests==2.31.0
reportlab==4.0.7
fpdf2==2.7.0
